JSObject updates

+++ SpawnDev.BlazorJS/JSObjects/ArrayBuffer.cs
@@ -1,8 +1,10 @@
 ﻿using Microsoft.JSInterop;
+using System.Text.Json.Serialization;
 
-namespace SpawnDev.BlazorJS.JSObjects {
+namespace SpawnDev.BlazorJS.JSObjects
+{
     public class ArrayBuffer : JSObject {
-        public int ByteLength => JSRef.Get<int>("byteLength");
+        public long ByteLength => JSRef.Get<long>("byteLength");
         public ArrayBuffer(long length) : base(JS.New(nameof(ArrayBuffer), length)) { }
         public ArrayBuffer(IJSInProcessObjectReference _ref) : base(_ref) { }
         public byte[] ReadBytes() {
